requirements.txt
Flask==2.0.1
Send2Trash==1.8.0
pytesseract==0.3.8
Pillow==8.3.1
